reimplementation for generative pipelines
