+++ src/deepsparse/pipeline.py
@@ -567,6 +567,34 @@ class Pipeline(ABC):
 
         return _register_pipeline_tasks_decorator
 
+    @staticmethod
+    def create_engine(
+        onnx_file_path: str,
+        engine_type: str,
+        engine_args: Dict,
+        context: Optional[Context] = None,
+    ) -> Union[Engine, MultiModelEngine, ORTEngine]:
+        engine_type = engine_type.lower()
+
+        if engine_type == DEEPSPARSE_ENGINE:
+            if context is not None and isinstance(context, Context):
+                engine_args.pop("num_cores", None)
+                engine_args.pop("scheduler", None)
+                engine_args["context"] = context
+                return MultiModelEngine(
+                    model=onnx_file_path,
+                    **engine_args,
+                )
+            return Engine(onnx_file_path, **engine_args)
+
+        if engine_type == ORT_ENGINE:
+            return ORTEngine(onnx_file_path, **engine_args)
+
+        raise ValueError(
+            f"Unknown engine_type {engine_type}. Supported values include: "
+            f"{SUPPORTED_PIPELINE_ENGINES}"
+        )
+
     @classmethod
     def from_config(
         cls,
@@ -791,26 +819,10 @@ class Pipeline(ABC):
         """
         return self.engine(engine_inputs)
 
-    def _initialize_engine(self) -> Union[Engine, ORTEngine]:
-        engine_type = self.engine_type.lower()
-
-        if engine_type == DEEPSPARSE_ENGINE:
-            if self.context is not None and isinstance(self.context, Context):
-                self._engine_args.pop("num_cores", None)
-                self._engine_args.pop("scheduler", None)
-                self._engine_args["context"] = self.context
-                return MultiModelEngine(
-                    model=self.onnx_file_path,
-                    **self._engine_args,
-                )
-            return Engine(self.onnx_file_path, **self._engine_args)
-        elif engine_type == ORT_ENGINE:
-            return ORTEngine(self.onnx_file_path, **self._engine_args)
-        else:
-            raise ValueError(
-                f"Unknown engine_type {self.engine_type}. Supported values include: "
-                f"{SUPPORTED_PIPELINE_ENGINES}"
-            )
+    def _initialize_engine(self) -> Union[Engine, MultiModelEngine, ORTEngine]:
+        return Pipeline.create_engine(
+            self.onnx_file_path, self.engine_type, self._engine_args, self.context
+        )
 
     def _identifier(self):
         # get pipeline identifier; used in the context of logging
